# Fix a bug that objects remain in storage.