I found a rare case bug that objects remain in a storage in spite of gc working.

We can sometimes see by the following.
```go
import (
	"fmt"
	"time"
	"strconv"
	"math/big"
	"github.com/OrlovEvgeny/go-mcache"
)

const (
	N = 4000
)

func test() {
	driver := mcache.StartInstance()

	var interval time.Duration
	var i int64

	fmt.Printf("Storage has %d object\n", driver.Len())

	i = 0
	for true {
		if i % 1000 == 0 {
			fmt.Printf("It is working... %d\n", i)
		}

		interval = time.Millisecond * 1

		time.Sleep(interval)

		driver.Set(strconv.FormatInt(i, 10), big.NewInt(i), interval)

		i++

		if i >= N {
			break
		}
	}

	fmt.Printf("Finished setting %d objects\n", N)

	fmt.Printf("Storage currently has %d object\n", driver.Len())

	time.Sleep(time.Second * 10)

	l := driver.Len()
	if l != 0 {
		fmt.Printf("%d objects remain\n", l)
	} else {
		fmt.Println("Nothing remain")
	}
}

func main() {
	test()
}
```

Please confirm the behavior above, and check my code.

---

**Performance**

Before
```
BenchmarkWrite-4          200000              8370 ns/op                                                                                                                                                    
BenchmarkRead-4          1000000              1710 ns/op                                                                                                                                                    
BenchmarkRW-4             300000              9412 ns/op
```

After
```
BenchmarkWrite-4          200000              7991 ns/op 
BenchmarkRead-4          1000000              1716 ns/op 
BenchmarkRW-4             300000              9894 ns/op
```